backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
sqlalchemy==2.0.25
# asyncpg==0.29.0  # Build failed locally, switching to psycopg v3 binary
psycopg[binary]>=3.2.0
redis==5.0.1
gitpython==3.1.41
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.13.1
tree-sitter>=0.21.0

